t0061: run_command executes scripts without a #! line
Add a test to 't0061-run-command.sh' to ensure that run_command can
continue to execute scripts which don't include a '#!' line.

As shell scripts are not natively executable on Windows, we use a
workaround to check "#!" when running scripts from Git.  As this
test requires the platform (not with Git's help) to run scripts
without "#!", skipt it on Windows.

Signed-off-by: Brandon Williams <bmwill@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>